Tidied up and simplified classes to comply with php mess detection rules
